// Neptune.Web/src/app/shared/generated/model/field-visit-provisional-grid-dto.ts
export class FieldVisitProvisionalGridDto { 
    FieldVisitID?: number;
    TreatmentBMPID?: number;
    TreatmentBMPName?: string | null;
    VisitDate?: string;
    StormwaterJurisdictionID?: number;
    StormwaterJurisdictionName?: string | null;
    PerformedByPersonID?: number;
    PerformedByPersonName?: string | null;
    FieldVisitStatusID?: number;
    FieldVisitStatusDisplayName?: string | null;
    FieldVisitTypeID?: number;
    FieldVisitTypeDisplayName?: string | null;
    IsFieldVisitVerified?: boolean;
    TreatmentBMPAssessmentIDInitial?: number | null;
    IsAssessmentCompleteInitial?: boolean;
    AssessmentScoreInitial?: number | null;
    MaintenanceRecordID?: number | null;
    TreatmentBMPAssessmentIDPM?: number | null;
    IsAssessmentCompletePM?: boolean;
    AssessmentScorePM?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}